moved exception classes into their own subdirectory
